requirements.txt
gradio
youtube-transcript-api
python-dotenv
google-generativeai
langchain
langchain-google-genai